Organize config tool into hostos and setupos
