+++ rs/ic_os/config/src/hostos/guestos_bootstrap_image.rs
