+++ rs/ic_os/config/src/hostos/mod.rs
@@ -1,0 +1,2 @@
+pub mod guestos_bootstrap_image;
+pub mod guestos_config;
