+++ rs/ic_os/config/src/lib.rs
@@ -1,8 +1,6 @@
-pub mod config_ini;
-pub mod deployment_json;
 pub mod generate_testnet_config;
-pub mod guestos_bootstrap_image;
-pub mod guestos_config;
+pub mod hostos;
+pub mod setupos;
 
 use anyhow::{Context, Result};
 use serde::{Deserialize, Serialize};
