+++ rs/ic_os/config/src/setupos/mod.rs
@@ -1,0 +1,2 @@
+pub mod config_ini;
+pub mod deployment_json;
