+++ rs/ic_os/dev_test_tools/launch-single-vm/src/main.rs
@@ -2,7 +2,7 @@ use clap::Parser;
 use config::generate_testnet_config::{
     generate_testnet_config, GenerateTestnetConfigArgs, Ipv6ConfigType,
 };
-use config::guestos_bootstrap_image::BootstrapOptions;
+use config::hostos::guestos_bootstrap_image::BootstrapOptions;
 use config_types::DeploymentEnvironment;
 use ic_prep_lib::{
     internet_computer::{IcConfig, TopologyConfig},
