+++ rs/tests/driver/src/driver/bootstrap.rs
@@ -28,7 +28,7 @@ use anyhow::{bail, Context, Result};
 use config::generate_testnet_config::{
     generate_testnet_config, GenerateTestnetConfigArgs, Ipv6ConfigType,
 };
-use config::guestos_bootstrap_image::BootstrapOptions;
+use config::hostos::guestos_bootstrap_image::BootstrapOptions;
 use config_types::DeploymentEnvironment;
 use ic_base_types::NodeId;
 use ic_prep_lib::{
